requirements.txt
flask
reportlab
google.generativeai